% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_fips}
\alias{state_fips}
\title{State FIPS codes.}
\format{A data.frame with 57 rows and 2 variables:
\describe{
  \item{st}{State abbreviation.}
  \item{fips}{state fips code.}
}}
\source{
U.S. Census Bureau, available at \url{https://www.census.gov/geo/reference/codes/cou.html}.
}
\usage{
state_fips
}
\description{
A data.frame containing the fips code and two-letter abbreviation for
states and territories in the U.S. See
\url{https://www.census.gov/geo/reference/county-changes.html} for changes
over time.
}
\keyword{datasets}

